Exclusão da classe gerenciar jogadores e adição de seus métodos na classe Tabuleiro
